Renames MoveAtom PositionerAtom
MoveAtom was too specific to motors. Positioner is more in keeping with
with other uses of the term in the scanning project.
