+++ org.eclipse.scanning.api/src/org/eclipse/scanning/api/event/queues/beans/PositionerAtom.java
