+++ org.eclipse.scanning.points/src/org/eclipse/scanning/points/classregistry/ScanningAPIClassRegistry.java
@@ -23,7 +23,7 @@ import org.eclipse.scanning.api.event.alive.HeartbeatBean;
 import org.eclipse.scanning.api.event.alive.KillBean;
 import org.eclipse.scanning.api.event.alive.PauseBean;
 import org.eclipse.scanning.api.event.queues.beans.MonitorAtom;
-import org.eclipse.scanning.api.event.queues.beans.MoveAtom;
+import org.eclipse.scanning.api.event.queues.beans.PositionerAtom;
 import org.eclipse.scanning.api.event.queues.beans.ScanAtom;
 import org.eclipse.scanning.api.event.queues.beans.SubTaskAtom;
 import org.eclipse.scanning.api.event.queues.beans.TaskBean;
@@ -158,7 +158,7 @@ public class ScanningAPIClassRegistry implements IClassRegistry {
 		// event.queues.beans
 		registerClass(tmp, QueueRequest.class);
 		registerClass(tmp, MonitorAtom.class);
-		registerClass(tmp, MoveAtom.class);
+		registerClass(tmp, PositionerAtom.class);
 		registerClass(tmp, ScanAtom.class);
 		registerClass(tmp, SubTaskAtom.class);
 		registerClass(tmp, TaskBean.class);
